Move the reference transaction enabled check into helpers.
With the 1st step we will move the class into helpers. The next step would be to refactor/rename it.
